python/requirements.txt
ollama
pandas==2.1.4
numpy==1.26.4
langchain
python-dotenv
pinecone-client
tiktoken
opencv-python==4.8.1.78
pytesseract
flask
flask-cors
pypdf
pdfplumber
python-docx==1.1.2
nltk
langdetect
google-generativeai
docling
python-doctr[torch] 
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
pdf2image